% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbn.R
\name{bbn}
\alias{bbn}
\title{Personal greeting}
\usage{
bbn()
}
\arguments{
\item{name}{Your name (character string; e.g. "john doe").}
}
\value{
A character string, capitalized to title case.
}
\description{
Greet a person and appropriately capitalize their name.
}
\examples{
hello("james bond")
}
